Set no wrapper ChatQnA as default
