+++ ChatQnA/Dockerfile.without_rerank
@@ -8,6 +8,7 @@ FROM python:3.11-slim
 RUN apt-get update -y && apt-get install -y --no-install-recommends --fix-missing \
     libgl1-mesa-glx \
     libjemalloc-dev \
+    vim \
     git
 
 RUN useradd -m -s /bin/bash user && \
@@ -19,9 +20,10 @@ RUN git clone https://github.com/opea-project/GenAIComps.git
 
 WORKDIR /home/user/GenAIComps
 RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r /home/user/GenAIComps/requirements.txt
+    pip install --no-cache-dir -r /home/user/GenAIComps/requirements.txt && \
+    pip install --no-cache-dir langchain_core
 
-COPY ./chatqna_without_rerank.py /home/user/chatqna_without_rerank.py
+COPY ./chatqna_no_wrapper.py /home/user/chatqna_no_wrapper.py
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user/GenAIComps
 
@@ -29,6 +31,4 @@ USER user
 
 WORKDIR /home/user
 
-RUN echo 'ulimit -S -n 999999' >> ~/.bashrc
-
-ENTRYPOINT ["python", "chatqna_without_rerank.py"]
+ENTRYPOINT ["python", "chatqna.py", "--without-rerank"]
