+++ ChatQnA/chatqna.yaml
@@ -30,21 +30,11 @@ opea_micro_services:
       HABANA_VISIBLE_DEVICES: all
       OMPI_MCA_btl_vader_single_copy_mechanism: none
     model-id: ${EMBEDDING_MODEL_ID}
-  embedding:
-    host: ${EMBEDDING_SERVICE_HOST_IP}
-    ports: ${EMBEDDING_SERVICE_PORT}
-    image: opea/embedding-tei:latest
-    endpoint: /v1/embeddings
   retrieval:
     host: ${RETRIEVER_SERVICE_HOST_IP}
     ports: ${RETRIEVER_SERVICE_PORT}
     image: opea/retriever-redis:latest
     endpoint: /v1/retrieval
-  reranking:
-    host: ${RERANK_SERVICE_HOST_IP}
-    ports: ${RERANK_SERVICE_PORT}
-    image: opea/reranking-tei:latest
-    endpoint: /v1/reranking
   tgi-service:
     host: ${TGI_SERVICE_IP}
     ports: ${TGI_SERVICE_PORT}
@@ -64,11 +54,6 @@ opea_micro_services:
       USE_FLASH_ATTENTION: true
       FLASH_ATTENTION_RECOMPUTE: true
     model-id: ${LLM_MODEL_ID}
-  llm:
-    host: ${LLM_SERVICE_HOST_IP}
-    ports: ${LLM_SERVICE_PORT}
-    image: opea/llm-tgi:latest
-    endpoint: /v1/chat/completions
   ui:
     host: ${UI_SERVICE_HOST_IP}
     ports:
